AITestEngine/requirements.txt
PyYAML
Django
django-sslserver
djangorestframework
django-rest-swagger
django-cors-headers
xlrd
mysql-connector-python
gensim
nltk
scikit-learn==0.20.2
cx-Oracle
gunicorn